% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtriang.R, R/ptriang.R, R/qtriang.R,
%   R/rtriang.R
\name{Triangular}
\alias{Triangular}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{The Triangular Distribution}
\usage{
dtriang(x, min = 0, max = 1, mode = (min + max)/2, log = FALSE)

ptriang(
  q,
  min = 0,
  max = 1,
  mode = (min + max)/2,
  lower.tail = TRUE,
  log.p = FALSE
)

qtriang(
  p,
  min = 0,
  max = 1,
  mode = (min + max)/2,
  lower.tail = TRUE,
  log.p = FALSE
)

rtriang(n, min = 0, max = 1, mode = (min + max)/2)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}

\item{mode}{mode of the distribution. Must be finite.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities p are given as
\code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
the length is taken to be the number required.}
}
\value{
\code{dtriang} gives the density function,
\code{ptriang} gives the distribution function,
\code{qtriang} gives the quantile function, and
\code{rtriang} generates random deviates.

The length of the result is determined by \code{n} for \code{rtriang}, and
is the maximum of the lengths of the numerical arguments for the other
functions.

TODO:
The numerical arguments other than \code{n} are recycled to the length of
the result.  Only the first elements of the logical arguments are used.
}
\description{
These functions provide information about the triangular
distribution on the interval from \code{min} to \code{max} with mode equal to
\code{mode}. \code{dtriang} gives the density function, \code{ptriang} gives
the distribution function, \code{qtriang} gives the quantile function, and
\code{rtriang} generates random deviates.
}
\details{
If \code{min}, \code{max}, or \code{mode} are not specified they assume the
default values of \code{0}, \code{1}, and \code{(min+max)/2}
respectively.

The triangular distribution has density
\deqn{0}{0}
for \eqn{x < min} or \eqn{x > max}
\deqn{f(x) = \frac{2(x-min)}{(max-min)(mode-min)}}{f(x) = 2(x-min)/(max-min)
  (mode-min)}
for \eqn{min \le x < mode}, and
\deqn{f(x) = \frac{2(max-x)}{(max-min)(max-mode)}}{E(x) = 2(max-x)/(max-min)
  (max-mode)}
for \eqn{mode < x \le max}.

TODO: \code{rtriang} will not generate either of the extreme values unless
\code{max == min} or \code{max-min} is small compared to \code{min}, and in
particular not for the default arguments.
}
\note{
The characteristics of output from pseudo-random number generators
(such as precision and periodicity) vary widely. See
\code{\link{.Random.seed}} for more information on R's random number
generation algorithms.
}
\examples{
 # TODO: Add examples

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
\emph{The New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{RNG}} about random number generation in \R.

\link{Distributions} for other standard distributions.
}
